+++ wpilibc/src/main/native/include/RobotState.h
@@ -11,23 +11,8 @@
 
 namespace frc {
 
-class RobotStateInterface {
- public:
-  virtual ~RobotStateInterface() = default;
-  virtual bool IsDisabled() const = 0;
-  virtual bool IsEnabled() const = 0;
-  virtual bool IsOperatorControl() const = 0;
-  virtual bool IsAutonomous() const = 0;
-  virtual bool IsTest() const = 0;
-};
-
 class RobotState {
- private:
-  static std::shared_ptr<RobotStateInterface> impl;
-
  public:
-  static void SetImplementation(RobotStateInterface& i);
-  static void SetImplementation(std::shared_ptr<RobotStateInterface> i);
   static bool IsDisabled();
   static bool IsEnabled();
   static bool IsOperatorControl();
