+++ wpilibcIntegrationTests/build.gradle
@@ -3,6 +3,7 @@ import org.gradle.language.base.internal.ProjectLayout
 apply plugin: 'cpp'
 apply plugin: 'visual-studio'
 apply plugin: 'edu.wpi.first.NativeUtils'
+apply plugin: ExtraTasks
 
 apply from: '../shared/config.gradle'
