+++ wpilibj/src/main/java/edu/wpi/first/wpilibj/RobotBase.java
@@ -21,8 +21,6 @@ import edu.wpi.first.wpilibj.hal.FRCNetComm.tInstances;
 import edu.wpi.first.wpilibj.hal.FRCNetComm.tResourceType;
 import edu.wpi.first.wpilibj.hal.HAL;
 import edu.wpi.first.wpilibj.hal.HALUtil;
-import edu.wpi.first.wpilibj.internal.HardwareHLUsageReporting;
-import edu.wpi.first.wpilibj.internal.HardwareTimer;
 import edu.wpi.first.wpilibj.livewindow.LiveWindow;
 import edu.wpi.first.wpilibj.util.WPILibVersion;
 
@@ -84,7 +82,6 @@ public abstract class RobotBase implements AutoCloseable {
    * to put this code into it's own task that loads on boot so ensure that it runs.
    */
   protected RobotBase() {
-    initializeHardwareConfiguration();
     NetworkTableInstance inst = NetworkTableInstance.getDefault();
     setupCameraServerShared();
     inst.setNetworkIdentity("Robot");
@@ -199,24 +196,6 @@ public abstract class RobotBase implements AutoCloseable {
   }
 
   /**
-   * Common initialization for all robot programs.
-   */
-  public static void initializeHardwareConfiguration() {
-    if (!HAL.initialize(500, 0)) {
-      throw new IllegalStateException("Failed to initialize. Terminating");
-    }
-
-    // Set some implementations so that the static methods work properly
-    Timer.SetImplementation(new HardwareTimer());
-    HLUsageReporting.SetImplementation(new HardwareHLUsageReporting());
-    RobotState.SetImplementation(DriverStation.getInstance());
-
-    // Call a CameraServer JNI function to force OpenCV native library loading
-    // Needed because all the OpenCV JNI functions don't have built in loading
-    CameraServerJNI.enumerateSinks();
-  }
-
-  /**
    * Starting point for the applications.
    */
   @SuppressWarnings({"PMD.AvoidInstantiatingObjectsInLoops", "PMD.AvoidCatchingThrowable",
@@ -225,6 +204,10 @@ public abstract class RobotBase implements AutoCloseable {
     if (!HAL.initialize(500, 0)) {
       throw new IllegalStateException("Failed to initialize. Terminating");
     }
+
+    // Call a CameraServer JNI function to force OpenCV native library loading
+    // Needed because all the OpenCV JNI functions don't have built in loading
+    CameraServerJNI.enumerateSinks();
 
     HAL.report(tResourceType.kResourceType_Language, tInstances.kLanguage_Java);
